add comment about com.newrelic.shared
Just adding this as an idea - at first I was not clear why the oTel example has a dependency on NR, until I looked deeper into the NR code. Possible that field team members may not be able to dig into the code quickly on this, while in a demo to customers. Hope is to prevent any customer devs from thinking that any NR-specific code is required, when instrumenting oTel and planning to ship to NR.